"Refactor UART import and add parse method to HSL_Frame class"
The code has been modified to replace the machine UART import with a custom class. This change was made to give more control over the UART properties. Additionally, a classmethod 'parse' has been added to the HSL_Frame class to provide a more efficient way to parse bytes. This could potentially improve the overall performance of the system.