"Fix typos, improve variable names and error messages, add to_bytes methods and README"
Corrected several typos across the ControlAndCommandWords.py file, notably changing 'UARTUrgrade' to 'UARTUpgrade' and 'InitializationCompletedQuerry' to 'InitializationCompletedQuery'. The variable name 'CNTW' was replaced with 'ControlWordBackpoint' for better code readability and convention compliance.

The Frame.py file was updated with a modification in class method and new to_bytes() method provided for both HSL_Frame and Frame classes.

Enhanced error handling in Errors.py by revamping the error messages to provide detailed context about the error for better debugging.

Added a README.md file, providing introductory information about the HumanStaticLitePy library, its installation, usage, and examples on how to parse and send data with the API.

A minor change was made in the __init__.py file.

These changes were made to improve the clarity and usability of the code.